pyproject.toml
[project]
name = "students-ai-assistance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.136.1",
    "ffmpeg-python>=0.2.0",
    "gtts>=2.5.4",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.9",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "pandas>=2.3.3",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "sounddevice>=0.5.5",
    "streamlit>=1.56.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.6.1",
    "uvicorn>=0.46.0",
]
